Implement Kronometer sessions
This feature replaces the previous XML 'save files' with Session objects.
All the session are saved in a global JSON file in the local application data
directory. The laps export in the XML format is replaced by the JSON format as
well. The KIO framework is no more required as dependency.