Improve usability of session dialog
Currently there is no prompt in the session dialog. One could add a label such
as "Select a session:", but that would look out of place. If we replace the OK
button with a 'Open session' button, then the user can realize what the dialog is
about: selecting a session and open it.