Replace QFrame with QGroupBox in TimeDisplay
TimeDisplay was basically reinventing the wheel, since a QFrame with a QLabel
as title is exactly a QGroupBox. The only drawback is that there is no easy way
to show/hide the title of the group box, but that was a setting that made
little sense anyway.

Plus, we now set the colors using QPalette instead of stile sheets, which is
the recommended way.