Improve toolbar look
Since there is no menubar anymore, we can reduce the icon size of the toolbar,
as there is no need anymore to "stand out".
We can also remove the "Control" text from the menu button. Its icon is already
auto-explicative (it's used also in Android) and the toolbar is already crowded
enough.