Do not assume sessions.json exists and is readable
sessions.json is created only when saving the session, so it does not
exist when opening from scratch, or also when using kronometer without
saving the session; in this case, QFile::open() fails (printing a
debug message about this), and the rest of the code just goes ahead
assuming everything went fine.

Instead, attempt to open the file only if it exists, and do the actual
loading only if the opening was successful (I/O errors may always
happen), logging the error otherwise.