Fix stopwatch restart after system sleep
m_accumulator was reset to 0 only inside the start() method if the
stopwatch was inactive. But in MainWindow::slotPrepareForSleep() we do
an unconditional pause() on the stopwatch, so this breaks the stopwatch
restart after the system resumes.

The proper fix is to reset m_accumulator inside the reset() method.

BUG: 439998
FIXED-IN: 2.3.0